Make the encoding of document IDs optional; necessary to allow reliable checking/creation of design documents.
Behaviour from Couch when performing a GET on a document ID with a url-encoded '/' (%2F) is to redirect to the same url with the unencoded version.  The get() method only looks for response code 200 and after ths redirect the response code is 301 (or, conceivably, 302), so it returns a null. Having tested for existence in this way and receiving a null response whether the document does exist or not, so a subsequent attempt to PUT to the same document ID fails if a document is already there.